+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/MessageUtils.kt
@@ -18,10 +18,11 @@ package io.getstream.chat.android.compose.ui.util
 
 import android.content.Context
 import com.getstream.sdk.chat.model.ModelType
-import com.getstream.sdk.chat.utils.EmojiUtil
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.compose.R
+import io.getstream.chat.android.uiutils.extension.isUploading
+import io.getstream.chat.android.uiutils.util.EmojiUtil
 
 /**
  * Takes the current message and returns the sender display name.